requirements.txt
gunicorn
pymongo
mongoengine
Flask
Faker
pytest
pyjwt
dnspython
psycopg2